+++ src/AksArc/AksArc.Autorest/exports/Update-AzAksArcCluster.ps1
@@ -39,10 +39,6 @@ Update-AzAksArcCluster -ClusterName azps_test_cluster -ResourceGroupName azps_te
 Update-AzAksArcCluster -ClusterName azps_test_cluster -ResourceGroupName azps_test_group -SmbCsiDriverEnabled:$false
 .Example
 Update-AzAksArcCluster -ClusterName azps_test_cluster -ResourceGroupName azps_test_group -AdminGroupObjectID @("2e00cb64-66d8-4c9c-92d8-6462caf99e33", "1b28ff4f-f7c5-4aaa-aa79-ba8b775ab443")
-.Example
-Update-AzAksArcCluster -ClusterName azps_test_cluster -ResourceGroupName azps_test_group -Upgrade
-.Example
-Update-AzAksArcCluster -ClusterName azps_test_cluster -ResourceGroupName azps_test_group -KubernetesVersion "1.28.5"
 
 .Outputs
 Microsoft.Azure.PowerShell.Cmdlets.AksArc.Models.IProvisionedCluster
@@ -119,24 +115,133 @@ param(
     # The default value is true.
     ${SmbCsiDriverEnabled},
 
+    [Parameter()]
+    [Microsoft.Azure.PowerShell.Cmdlets.AksArc.Category('Body')]
+    [System.String]
+    # Valid values are 'true' and 'false'
+    ${AutoScalerProfileBalanceSimilarNodeGroup},
+
+    [Parameter()]
+    [Microsoft.Azure.PowerShell.Cmdlets.AksArc.PSArgumentCompleterAttribute("least-waste", "most-pods", "priority", "random")]
+    [Microsoft.Azure.PowerShell.Cmdlets.AksArc.Category('Body')]
+    [System.String]
+    # If not specified, the default is 'random'.
+    # See [expanders](https://github.com/kubernetes/autoscaler/blob/master/cluster-autoscaler/FAQ.md#what-are-expanders) for more information.
+    ${AutoScalerProfileExpander},
+
+    [Parameter()]
+    [Microsoft.Azure.PowerShell.Cmdlets.AksArc.Category('Body')]
+    [System.String]
+    # The default is 10.
+    ${AutoScalerProfileMaxEmptyBulkDelete},
+
+    [Parameter()]
+    [Microsoft.Azure.PowerShell.Cmdlets.AksArc.Category('Body')]
+    [System.String]
+    # The default is 600.
+    ${AutoScalerProfileMaxGracefulTerminationSec},
+
+    [Parameter()]
+    [Microsoft.Azure.PowerShell.Cmdlets.AksArc.Category('Body')]
+    [System.String]
+    # The default is '15m'.
+    # Values must be an integer followed by an 'm'.
+    # No unit of time other than minutes (m) is supported.
+    ${AutoScalerProfileMaxNodeProvisionTime},
+
+    [Parameter()]
+    [Microsoft.Azure.PowerShell.Cmdlets.AksArc.Category('Body')]
+    [System.String]
+    # The default is 45.
+    # The maximum is 100 and the minimum is 0.
+    ${AutoScalerProfileMaxTotalUnreadyPercentage},
+
+    [Parameter()]
+    [Microsoft.Azure.PowerShell.Cmdlets.AksArc.Category('Body')]
+    [System.String]
+    # For scenarios like burst/batch scale where you don't want CA to act before the kubernetes scheduler could schedule all the pods, you can tell CA to ignore unscheduled pods before they're a certain age.
+    # The default is '0s'.
+    # Values must be an integer followed by a unit ('s' for seconds, 'm' for minutes, 'h' for hours, etc).
+    ${AutoScalerProfileNewPodScaleUpDelay},
+
+    [Parameter()]
+    [Microsoft.Azure.PowerShell.Cmdlets.AksArc.Category('Body')]
+    [System.String]
+    # This must be an integer.
+    # The default is 3.
+    ${AutoScalerProfileOkTotalUnreadyCount},
+
+    [Parameter()]
+    [Microsoft.Azure.PowerShell.Cmdlets.AksArc.Category('Body')]
+    [System.String]
+    # The default is '10m'.
+    # Values must be an integer followed by an 'm'.
+    # No unit of time other than minutes (m) is supported.
+    ${AutoScalerProfileScaleDownDelayAfterAdd},
+
+    [Parameter()]
+    [Microsoft.Azure.PowerShell.Cmdlets.AksArc.Category('Body')]
+    [System.String]
+    # The default is the scan-interval.
+    # Values must be an integer followed by an 'm'.
+    # No unit of time other than minutes (m) is supported.
+    ${AutoScalerProfileScaleDownDelayAfterDelete},
+
+    [Parameter()]
+    [Microsoft.Azure.PowerShell.Cmdlets.AksArc.Category('Body')]
+    [System.String]
+    # The default is '3m'.
+    # Values must be an integer followed by an 'm'.
+    # No unit of time other than minutes (m) is supported.
+    ${AutoScalerProfileScaleDownDelayAfterFailure},
+
+    [Parameter()]
+    [Microsoft.Azure.PowerShell.Cmdlets.AksArc.Category('Body')]
+    [System.String]
+    # The default is '10m'.
+    # Values must be an integer followed by an 'm'.
+    # No unit of time other than minutes (m) is supported.
+    ${AutoScalerProfileScaleDownUnneededTime},
+
+    [Parameter()]
+    [Microsoft.Azure.PowerShell.Cmdlets.AksArc.Category('Body')]
+    [System.String]
+    # The default is '20m'.
+    # Values must be an integer followed by an 'm'.
+    # No unit of time other than minutes (m) is supported.
+    ${AutoScalerProfileScaleDownUnreadyTime},
+
+    [Parameter()]
+    [Microsoft.Azure.PowerShell.Cmdlets.AksArc.Category('Body')]
+    [System.String]
+    # The default is '0.5'.
+    ${AutoScalerProfileScaleDownUtilizationThreshold},
+
+    [Parameter()]
+    [Microsoft.Azure.PowerShell.Cmdlets.AksArc.Category('Body')]
+    [System.String]
+    # The default is '10'.
+    # Values must be an integer number of seconds.
+    ${AutoScalerProfileScanInterval},
+
+    [Parameter()]
+    [Microsoft.Azure.PowerShell.Cmdlets.AksArc.Category('Body')]
+    [System.String]
+    # The default is true.
+    ${AutoScalerProfileSkipNodesWithLocalStorage},
+
+    [Parameter()]
+    [Microsoft.Azure.PowerShell.Cmdlets.AksArc.Category('Body')]
+    [System.String]
+    # The default is true.
+    ${AutoScalerProfileSkipNodesWithSystemPod},
+
     [Parameter(ParameterSetName='AutoScaling', Mandatory)]
     [Microsoft.Azure.PowerShell.Cmdlets.AksArc.Category('Body')]
     [System.Management.Automation.SwitchParameter]
     # Indicates whether to enable autoscalar.
     # The default value is true.
     ${EnableAutoScaling},
-
-    [Parameter(ParameterSetName='Upgrade', Mandatory)]
-    [Microsoft.Azure.PowerShell.Cmdlets.AksArc.Category('Body')]
-    [System.String]
-    # The version of Kubernetes in use by the provisioned cluster.
-    ${KubernetesVersion},
-
-    [Parameter(ParameterSetName='Upgrade2', Mandatory)]
-    [Microsoft.Azure.PowerShell.Cmdlets.AksArc.Category('Body')]
-    [System.Management.Automation.SwitchParameter]
-    # Upgrade the provisioned cluster
-    ${Upgrade},
 
     [Parameter()]
     [Microsoft.Azure.PowerShell.Cmdlets.AksArc.Category('Runtime')]
@@ -218,10 +323,8 @@ begin {
         $mapping = @{
             UpdateExpanded = 'Az.AksArc.custom\Update-AzAksArcCluster';
             AutoScaling = 'Az.AksArc.custom\Update-AzAksArcCluster';
-            Upgrade = 'Az.AksArc.custom\Update-AzAksArcCluster';
-            Upgrade2 = 'Az.AksArc.custom\Update-AzAksArcCluster';
         }
-        if (('UpdateExpanded', 'AutoScaling', 'Upgrade', 'Upgrade2') -contains $parameterSet -and -not $PSBoundParameters.ContainsKey('SubscriptionId') ) {
+        if (('UpdateExpanded', 'AutoScaling') -contains $parameterSet -and -not $PSBoundParameters.ContainsKey('SubscriptionId') ) {
             $testPlayback = $false
             $PSBoundParameters['HttpPipelinePrepend'] | Foreach-Object { if ($_) { $testPlayback = $testPlayback -or ('Microsoft.Azure.PowerShell.Cmdlets.AksArc.Runtime.PipelineMock' -eq $_.Target.GetType().FullName -and 'Playback' -eq $_.Target.Mode) } }
             if ($testPlayback) {