+++ src/NewRelic/NewRelic.Autorest/test/New-AzNewRelicMonitor.Tests.ps1
@@ -14,8 +14,22 @@ if(($null -eq $TestName) -or ($TestName -contains 'New-AzNewRelicMonitor'))
   . ($mockingPath | Select-Object -First 1).FullName
 }
 
-Describe 'New-AzNewRelicMonitor' {
-    It 'CreateExpanded' -skip {
+Describe 'New-AzNewRelicMonitor' -Tag 'LiveOnly' {
+    It 'CreateExpanded' {
+        {
+            New-AzResourceGroup -Name $env.resourceGroup -Location $env.region # $rgName = $rg.ResourceGroupName
+
+            $monitor = New-AzNewRelicMonitor -Name $env.NewMonitorName -ResourceGroupName $env.resourceGroup -Location $env.region -PlanDataPlanDetail $env.planDetails -PlanDataBillingCycle $env.billingCycle -PlanDataUsageType $env.usageType -PlanDataEffectiveDate (Get-Date -DisplayHint DateTime) -UserInfoEmailAddress $env.testerEmail -UserInfoFirstName $env.testerFirstName -UserInfoLastName $env.testerLastName
+            $monitor.Name | Should -Be $env.NewMonitorName
+            $monitor.ResourceGroup | Should -Be $env.resourceGroup
+        } | Should -Not -Throw
+    }
+
+    It 'CreateViaJsonFilePath' -skip {
+        { throw [System.NotImplementedException] } | Should -Not -Throw
+    }
+
+    It 'CreateViaJsonString' -skip {
         { throw [System.NotImplementedException] } | Should -Not -Throw
     }
 }