+++ src/Storage/Storage.Management/ChangeLog.md
@@ -18,6 +18,8 @@
         - Additional information about change #1
 -->
 ## Upcoming Release
+
+## Version 7.1.0
 * Fixed the issue that Azure.Core.AccessToken is used before assigned.
 * Supported TLS1_3 when creating and updating a storage account 
     - `New-AzStorageAccount`
@@ -25,6 +27,8 @@
 * Fixed sync copy blob issue with -AsJob is specified [#25105]
     - `Copy-AzStorageBlob`
 * Updated Storage.Management.Sdk to support API version 2023-05-01
+* Updated 2 help examples of create storage account cmdlet, with MinimumTlsVersion as TLS1_2.
+    - `New-AzStorageAccount`
 
 ## Version 7.0.0
 * Added a prompt that needs confirmation when upgrading a storage account from StorageV1 or BlobStorage to StorageV2. Can be suppressed with -Force.